[cython] enhance models.business API
Transfer to it some of the code that was done in the SolverModelBuilder class
in pytam (class responsible for translating the business model into a solver
model).

It is better to encapsulate this code since it will make the business to
solver model conversion code easier to understand. Besides it belongs to
the business model since it is strongly tight to the C++ code.

business2microsource is added in tympan.pytam module and used in
tympan.models.business module. It should be part of tympan.business2solver
module, and will be moved once the latter will have been extracted from pytam.


Related to #2015869