Fix solver access from cython library
Add a TYPlugin* in TYCalcul class and use it to retrieve the solver and the
plugin information in the go() method instead of doing it through the
TYPluginManager.

This is because now we run several cython libraries, each one loads its own
copy of tympan libraries and the OGenID of the solver in the TYPluginManager
of one copy isn't the same in the TYPluginManager of the other one.