[altimetry] Change test to reveal an orientaion bug and fix it
The CGAL method vertices_in_constraint does NOT necessarily return
the vertices in the same orientation as the constraint.

This went unnoticed in the test becaus it was not the case ; thus
the orientation of the constraint is changed to also reveal this
pitfall.

Fix the bug by explicitely checking - and correcting - the
orientation of yielded subconstraints with respect to the input
constraint.