[cython] Reinject solver results into business model
This code will replace the processing currently done in TYCalcul/TYResultat.
In this changeset, the processing is kept in these business classes but
slightly adapt for it to work with the new solver version.

This code aims at aggregating all the results into a reduced matrix, and send back
the updated values to the business model.

XXX: for now, it segfaults with some XML projects

Related to #2741289