[altimetry] Ignore isolated points in datamodel elementary shapes
The computation of the intersection between the site landtake and a levelcurve
(computed in merge.SiteNodeGeometryCleaner.process_level_curves)
can lead to the generation of a GeometryCollection containing some points
(see for example tests/data/projects-panel/TEST_DIFFRACTION_BUTTE_CARTO_NO_RESU.xml).

This shouldn't raise an exception. For now, just warn the user.

Related to #2741286