[cython] write TYResultat result matrix in DB and not in a linear scale
To do so, it is necessary to explicitely set the spectra states as SPECTRE_ETAT_LIN
(linear) before calling toDB().
As a matter of fact, the solver outputs the results in a linear scale,
but when a new spectra is built, its default state is set to SPECTRE_ETAT_DB (log
scale) no matter what it contains.

Related to #2741289