[altimetry] remove call to TYSiteNode::updateAltimetrie from TYSiteNode::update
As a matter of fact, update() is a recursive method. It allows to update the
altitude of the infrastructure elements and of the receptors, and to update
the acoustic for a site and for all is subsites.
However, the altimetry is now computed once, on the root site, for all the
included subsites. This call can't be recursive.

Related to #2741286