[altimetry] Drop clear_cache methods from mesher classes methods
This whole *cache* thing is a bad design idea and it has significant (sometime
huge) performance costs in some cases.

In `insert_constraint`, the only thing that's apparently updated is
the `_input_constraints_infos` attribute which is not affected by
`clear_caches` so it should not change anything here not to call
`clear_caches`.

In `insert_point`, the `_input_vertices_infos` is first updated in
`MeshedCDTWithInfo` class and the `vertices_infos` is then updated from the
latter (through the `input_vertex_infos` accessor method). So here,
`clear_caches` will first reconstruct the `vertices_infos` dict identical to
its prior state and `insert_point` will only add a new entry to both the
"private" and the public attribute. Which let's one wonder why there are two
attributes instead of just one.

Drop the `_init_vertices_info_from_input` along the way as its call in
`__init__` is now a no-op.

Closes #2809853.