Give the merged site to MeshFiller instead of just the main one
Some infrastructure elements may be defined out of the main site, and giving only the main site
lead to an error because these elements were not found there.

The test added used to crash before this correction. Having a building defined outside a site,
or intersecting it shouldn't lead to an error. However, the building should not be part of the
equivalent site, since it doesn't make much sense.