Add altimetry road script
Script for launching a Tympan calculation, taking as input parameters:
- an XML Tympan model,
- a CSV road file or a directory that contains one or several CSV road
files.
- the target directory

First, the script loads the project and builds the site.
Then, he creates and adds the road(s) and reconstructs the altimetry.
After updating the project site to take into account the roads,
it saves the mesh (PLY), calls the solver and saves results and
project.